extractd out receive swap handling logic
