+++ lib/core/src/lib.rs
@@ -7,6 +7,7 @@ pub mod frb;
 pub mod logger;
 pub mod model;
 pub mod persist;
+pub(crate) mod receive_swap;
 pub mod sdk;
 pub(crate) mod send_swap;
 pub(crate) mod swapper;
